.: Fixed: size and position of GUI controls in Shared Files page .: Changed: EncryptedDatagramSocket without buffer reallocation .: Changed: for completed files use part count as available parts number instead of zero .: Changed: ReplaceVistaIcon and EditBinX removed from build .: A few code edits; mostly for stricter typing and for getting smaller code size

+++ ClientCredits.h
@@ -79,7 +79,7 @@ public:
 	uint64	GetUploadedTotal() const;
 	uint64	GetDownloadedTotal() const;
 	float	GetScoreRatio(uint32 dwForIP) const;
-	void	SetLastSeen()					{m_pCredits->nLastSeen = time(NULL);}
+	void	SetLastSeen()						{m_pCredits->nLastSeen = static_cast<uint32>(time(NULL));}
 	bool	SetSecureIdent(const uchar* pachIdent, uint8 nIdentLen); // Public key cannot change, use only if there is not public key yet
 	uint32	m_dwCryptRndChallengeFor;
 	uint32	m_dwCryptRndChallengeFrom;
